Made CNI isolator cleanup more robust.
If the container is destroyed while in isolator preparing state, the
cleanup might fail due to missing files or directories. This patch makes
the cleanup path in CNI isolator more robust so that the cleanup does
not fail in those scenarios.

Review: https://reviews.apache.org/r/68333
(cherry picked from commit 3c79314e24592b6bd82457249746500d27c4e072)

+++ src/slave/containerizer/mesos/isolators/network/cni/cni.cpp
@@ -1545,14 +1545,16 @@ Future<Nothing> NetworkCniIsolatorProcess::_cleanup(
               << target << "' for container " << containerId;
   }
 
-  Try<Nothing> rmdir = os::rmdir(containerDir);
-  if (rmdir.isError()) {
-    return Failure(
-        "Failed to remove the container directory '" +
-        containerDir + "': " + rmdir.error());
-  }
+  if (os::exists(containerDir)) {
+    Try<Nothing> rmdir = os::rmdir(containerDir);
+    if (rmdir.isError()) {
+      return Failure(
+          "Failed to remove the container directory '" +
+          containerDir + "': " + rmdir.error());
+    }
 
-  LOG(INFO) << "Removed the container directory '" << containerDir << "'";
+    LOG(INFO) << "Removed the container directory '" << containerDir << "'";
+  }
 
   infos.erase(containerId);
 
@@ -1595,6 +1597,22 @@ Future<Nothing> NetworkCniIsolatorProcess::detach(
       rootDir.get(),
       containerId,
       networkName);
+
+  // There are two cases that the network config file might not exist
+  // when `detach` happens:
+  // (1) The container is destroyed when preparing. In that case, we
+  //     know that `attach` hasn't been called. Therefore, no need to
+  //     call `detach`.
+  // (2) The agent crashes when isolating, but before the network
+  //     config file is checkpointed. In that case, we also know that
+  //     CNI ADD command hasn't been called. Therefore, no need to
+  //     call `detach`.
+  if (!os::exists(networkConfigPath)) {
+    LOG(WARNING) << "Skip detach since network config file for container "
+                 << containerId << " and network name '" << networkName << "' "
+                 << "does not exist";
+    return Nothing();
+  }
 
   Try<JSON::Object> networkConfigJSON = getNetworkConfigJSON(
       networkName,
